Fixed header files

+++ Marcy/src/Torus.c
@@ -16,13 +16,13 @@
 
 int createTorus(int size)
 {
-	unsigned long long length = (1 << size);
-	unsigned long long uniqueCells = length * length;
+	unsigned long length = (1 << size);
+	unsigned long uniqueCells = length * length;
 
 	//struct Cells **allTheCells = malloc(uniqueCells * sizeof(struct Cells));		//This array will store the memloc to each cell
 
 	//First allocate memory for each cell
-	unsigned long long i = 0;
+	unsigned long i = 0;
 	for(i = 0; i < uniqueCells; i++)
 	{
 		allTheCells[i] = malloc(sizeof(struct Cells));
